// src/resources/Add_adressUserInterface.ts
export enum Add_adressUserInterface {

    buttom_account = '//*[@id="top-links"]/ul/li[2]/a',
    buttom_my_account = '//*[@id="top-links"]/ul/li[2]/ul/li[1]/a',
    buttom_add_adress = '//*[@id="content"]/ul[1]/li[3]/a',
    buttom_new_adress = '//*[@id="content"]/div/div[2]/a',
    first_name = '//*[@id="input-firstname"]',
    last_name = '//*[@id="input-lastname"]',
    company = '//*[@id="input-company"]',
    address1 = '//*[@id="input-address-1"]',
    address2 = '//*[@id="input-address-2"]',
    city = '//*[@id="input-city"]',
    post_code = '//*[@id="input-postcode"]',
    country = '//*[@id="input-country"]',
    region_State = '//*[@id="input-zone"]',
    buttom_continue = '//*[@id="content"]/form/div/div[2]/input'

}